enhancements and refactoring

+++ archivy/data.py
@@ -131,6 +131,9 @@ def get_item(dataobj_id):
         data = frontmatter.load(file)
         data["fullpath"] = str(file)
         data["dir"] = str(file.parent.relative_to(get_data_dir()))
+        # replace . for root items to ''
+        if data["dir"] == ".":
+            data["dir"] = ""
         return data
     return None
 
@@ -138,7 +141,13 @@ def get_item(dataobj_id):
 def move_item(dataobj_id, new_path):
     """Move dataobj of given id to new_path"""
     file = get_by_id(dataobj_id)
-    return shutil.move(file, f"{get_data_dir()}/{new_path}/")
+    data_dir = get_data_dir()
+    out_dir = data_dir / new_path
+    if (out_dir / file.parts[-1]).exists():
+        raise FileExistsError
+    elif is_relative_to(out_dir, data_dir) and out_dir.exists():  # check file isn't
+        return shutil.move(file, f"{get_data_dir()}/{new_path}/")
+    return False
 
 
 def delete_item(dataobj_id):
@@ -242,8 +251,6 @@ def get_dirs():
         if dir_path.is_dir()
     ]
 
-    # append name for root dir
-    dirnames.append("not classified")
     return dirnames
